+++ packages/vue-dot/playground/Playground.vue
@@ -50,13 +50,13 @@
 
 				<FormSectionEx />
 
+				<FormBuilderEx />
+
 				<LangBtnEx />
 
 				<NotificationEx />
 
 				<PaginatedTableEx />
-
-				<FormBuilderEx />
 
 				<RulesEx />
