+++ packages/vue-dot/src/patterns/FormSection/tests/data/questionForm.ts
@@ -167,10 +167,11 @@ export const questionForm: FieldGroup = {
 	},
 	questionChoiceButton: {
 		type: 'select',
-		title: 'Question choix bouton',
+		title: 'Question choix bouton inline',
 		description: 'Informations supplémentaires',
 		items: defaultItems,
 		metadata: {
+			inline: true,
 			type: 'choiceButton',
 			hint: 'Texte informatif'
 		}
